Add file assessment1.py
this python code creates a connection object to an SQLite database and a cursor object to execute SQL commands. Prompt the user for a file name or set the default file name for a list of email addresses. Then it splits, removes spaces, and counts the number of organizations in the email file. Finally, the code Queries the Counts table for the top 10 organizations with the most email messages and print the results to a new database if it does not exist yet.